refactored vue for domain driven arch
